Added a new struct to network.h, and changed x64_AES.s be an assembly file with preproccessor directives
Added a new struct, Packet_Header_t, to network.h, meant to be a header for packets. Changed the file ending of x64_AES.s from .s, an assembly source file without preproccessor directives, to .S, an assembly source file with preproccessor directives, so that we can detect if we are compiling for x64, and remove the code accordingly.